New InverseSum representation
An array is always a sum, and always an inverse of that sum
except at the top level.

I am using Javascript generator functions, and they are operating
as designed, but I am still working out how to narrow the range
of expressions at the top level.

I added the ability to provide an expression JSON string as an `expr`
query parameter in the URL.